Move SearchView to file in search directory
